+++ opentelemetry-sdk/src/metrics/internal/mod.rs
@@ -5,7 +5,6 @@ mod last_value;
 mod sum;
 
 use core::fmt;
-use opentelemetry_api::metrics::Result;
 use std::ops::{Add, AddAssign, Sub};
 
 pub(crate) use aggregator::Aggregator;
@@ -20,6 +19,7 @@ pub(crate) trait Number<T>:
     Add<Output = T>
     + AddAssign
     + Sub<Output = T>
+    + PartialOrd
     + fmt::Debug
     + Clone
     + Copy
@@ -29,23 +29,23 @@ pub(crate) trait Number<T>:
     + Sync
     + 'static
 {
-    fn try_into_float(&self) -> Result<f64>;
+    fn into_float(self) -> f64;
 }
 
 impl Number<i64> for i64 {
-    fn try_into_float(&self) -> Result<f64> {
+    fn into_float(self) -> f64 {
         // May have precision loss at high values
-        Ok(*self as f64)
+        self as f64
     }
 }
 impl Number<u64> for u64 {
-    fn try_into_float(&self) -> Result<f64> {
+    fn into_float(self) -> f64 {
         // May have precision loss at high values
-        Ok(*self as f64)
+        self as f64
     }
 }
 impl Number<f64> for f64 {
-    fn try_into_float(&self) -> Result<f64> {
-        Ok(*self)
+    fn into_float(self) -> f64 {
+        self
     }
 }